code fix ver.2.1
