+++ Epam/Java Basics_UpSkill Lab #1/Algorithmization/Sorting/src/com/androsor/sort/ArrayCreator.java
@@ -1,19 +1,13 @@
 package com.androsor.sort;
 
-import java.util.Arrays;
 import java.util.Random;
-import java.util.Scanner;
 
-import static java.lang.Math.abs;
-
+/**
+ * Creates an array and fills it with random numbers.
+ */
 public class ArrayCreator {
 
-    public static int inputParameter() {
-        Scanner scanner = new Scanner(System.in);
-        return abs(scanner.nextInt());
-    }
-
-    public static int[] createArray(int length) {
+    public static int[] createArrayRandom(int length) {
         int[] numbers = new int[length];
         Random random = new Random();
         for (int i = 0; i < length; i++) {
@@ -21,11 +15,4 @@ public class ArrayCreator {
         }
         return numbers;
     }
-
-    public static void printArray(int[] array) {
-        System.out.println("----------------------------------------");
-        System.out.println(" " + Arrays.toString(array));
-        System.out.println("----------------------------------------");
-    }
 }
-
