+++ Epam/Java Basics_UpSkill Lab #1/Algorithmization/Sorting/src/com/androsor/sort/HomeTask2.java
@@ -1,10 +1,9 @@
 package com.androsor.sort;
 
-import java.util.Arrays;
-
-import static com.androsor.sort.ArrayCreator.inputParameter;
-import static com.androsor.sort.ArrayCreator.createArray;
-import static com.androsor.sort.ArrayCreator.printArray;
+import static com.androsor.sort.ArrayCreator.createArrayRandom;
+import static com.androsor.sort.IOUtils.enterParameterFromConsole;
+import static com.androsor.sort.IOUtils.printArray;
+import static java.util.Arrays.copyOf;
 
 /**
  * Implement selection sort.
@@ -14,10 +13,10 @@ public class HomeTask2 {
     public static void main(String[] args) {
 
         System.out.print(" Введите количество элементов массива N = ");
-        int length = inputParameter();
+        int length = enterParameterFromConsole();
 
         System.out.println(" Исходный массив с количеством элементов N = " + length);
-        int[] numbers = createArray(length);
+        int[] numbers = createArrayRandom(length);
         printArray(numbers);
 
         System.out.println(" Отсортированный массив по возрастанию ");
@@ -30,7 +29,7 @@ public class HomeTask2 {
     }
 
     private static int[] sortAscending(int[] numbers) {
-        int[] copyArray = Arrays.copyOf(numbers, numbers.length);
+        int[] copyArray = copyOf(numbers, numbers.length);
         for (int i = 0; i < copyArray.length; i++) {
             int minItem = copyArray[i];
             int indexMinItem = i;
@@ -46,7 +45,7 @@ public class HomeTask2 {
     }
 
     private static int[] sortDescending(int[] numbers) {
-        int[] copyArray = Arrays.copyOf(numbers, numbers.length);
+        int[] copyArray = copyOf(numbers, numbers.length);
         for (int i = 0; i < copyArray.length; i++) {
             int maxItem = copyArray[i];
             int indexMaxItem = i;
